Add underlines back to links
The link color is so close to the text color that it is a bit
difficult to notice links. To make the links more readable, we can
simply add underlines back to the links.

It looks like the original version of the site also had a
.pretty-links class, which was only used on some of the pages. This
doesn't seem to really be necessary, so I simply removed it.

Special care was taken to avoid underlining links in the nav bar.